+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 2.0.2
-- Introduced add-ons management that allow installation of add-ons for various purposes. 
+- Introduced add-ons management that allow installation of add-ons for various purposes;
+- Added option for playlist block to set alignment;
+- Optimized resizing for playlist block theme Default;
+- Added option for video position block to set alignment;
+- Added option for single video embed to set alignment.
 
 2.0.1 - [14/Sep/2020]
 - Solved a bug that wasn't hiding the video if video was published as block element and filter to prevent the video embed was on;